Move check cast visitor to a separate file
