+++ runtime/sema/checker.go
@@ -115,6 +115,7 @@ type Checker struct {
 	checkHandler                       CheckHandlerFunc
 	expectedType                       Type
 	memberAccountAccessHandler         MemberAccountAccessHandlerFunc
+	lintEnabled                        bool
 }
 
 type Option func(*Checker) error
@@ -231,6 +232,16 @@ func WithPositionInfoEnabled(enabled bool) Option {
 			checker.FunctionInvocations = NewFunctionInvocations()
 		}
 
+		return nil
+	}
+}
+
+// WithLintingEnabled returns a checker option which enables/disables
+// advanced linting.
+//
+func WithLintingEnabled(enabled bool) Option {
+	return func(checker *Checker) error {
+		checker.lintEnabled = enabled
 		return nil
 	}
 }
